requirements.txt
asgiref==3.2.10
click==7.1.2
cmake==3.17.3
Django==3.0.14
django-filter==2.3.0
# dlib==19.22.0.\
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pytz==2020.1
sqlparse==0.3.1
